Move component to where it is used
